Reorganize data files
First off, delete a number of python files that were part of a `data` directory in src/ITR that actually do the data handling for ITR, and are not, themselves, data that can be loaded.

Secondly, reference the newly created `data_dir` path, which provides the default benchmark data for running ITR.  Of course, users can make their own benchmarks and load those.  But if they want to use the defaults, they now come from `ITR.data_dir`.

Thirdly, create a `data` directory that can be shared between `src/ITR_examples` and `test`, and use that for data that the example programs might need.

Finally, conform the use cases to reference these new structures.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>